💄 style(imports): consolidate import paths for make_bar
- use dedicated import path for make_bar from utils.misc across multiple modules
- enhance code readability and maintainability by centralizing imports
