+++ src/argoproxy/endpoints/chat.py
@@ -20,10 +20,10 @@ from ..types import (
     NonStreamChoice,
     StreamChoice,
 )
+from ..utils.misc import make_bar
 from ..utils.models import resolve_model_name
 from ..utils.tokens import calculate_prompt_tokens, count_tokens
 from ..utils.transports import send_off_sse
-from ..utils.utils import make_bar
 
 DEFAULT_MODEL = "gpt4o"
